backend/pyproject.toml
[tool.poetry]
name = "candidate-management-platform"
version = "0.1.0"
description = ""
authors = ["augustojulio <augusto.juliogec@gmail.com>"]
readme = "README.md"
packages = [{include = "candidate_management_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
black = "^22.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
